Crate oak_sign

Structs

KeyPair
Convenience struct that encapsulates ring::signature::Ed25519KeyPair.

SignatureBundle

Constants

HASH_TAG

PRIVATE_KEY_TAG

PUBLIC_KEY_TAG

SIGNATURE_TAG

Functions

get_sha256
Computes a SHA-256 digest of input and returns it in a form of raw bytes.

get_sha256_hex
Computes a SHA-256 digest of bytes and returns it in a hex encoded string.

read_pem_file

write_pem_file